+++ pyincore/analyses/cumulativebuildingdamage/cumulativebuildingdamage.py
@@ -14,7 +14,10 @@ from pyincore import BaseAnalysis, AnalysisUtil
 
 
 class CumulativeBuildingDamage(BaseAnalysis):
-    """Cumulative building damage analysis."""
+    """This analysis computes building damage based on a particular hazard. The process for computing the structural
+     damage is done externally and the results for earthquake and tsunami are imported to the analysis.
+     The damage intervals are then calculated from combined limit states.
+    """
 
     def run(self):
         """Executes Cumulative Building Damage Analysis"""