+++ pyincore/analyses/capitalshocks/__init__.py
@@ -1,0 +1,1 @@
+from pyincore.analyses.capitalshocks.capitalshocks import CapitalShocks