+++ tests/test_format.py
@@ -35,7 +35,12 @@ paths = [
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/montecarlofailureprobability/'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/dataset.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/hazardservice.py'),
-    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_hazardservice.py')
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_hazardservice.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/capitalshocks/capitalshocks.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/capitalshocks/test_capitalshocks.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_hazardservice.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/joplincgeanalysis/test_joplincgeanalysis.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/utils/analysisutil.py'),
 ]